backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
python-dotenv
groq
httpx
openai
pydantic
pydantic-settings
ruff
requests
